fix: build with mjs/cjs extensions
This was previously only tested with Nodejs 22, which worked fine,
however with lower versions, node is unable to determine what format the
bundle is just from the conditional exports.

To help these versions, we can generate esm/cjs modules with mjs/cjs, so
that node knows which format to load the files as.